Authentication refactoring
