+++ video.kino.pub/resources/lib/authwindow.py
@@ -1,15 +1,50 @@
 # -*- coding: utf-8 -*-
-
 import json
+import sys
 import time
-import threading
 import xbmc
 import xbmcgui
 import urllib
 import urllib2
+from addonutils import update_device_info
+from data import __addon__, __plugin__, __settings__, __skinsdir__
+
+_ADDON_PATH = xbmc.translatePath(__addon__.getAddonInfo("path"))
+
+if sys.platform in ("win32", "win64"):
+    _ADDON_PATH = _ADDON_PATH.decode("utf-8")
+
+
+class AuthWindow(xbmcgui.WindowXMLDialog):
+    def __init__(self, *args, **kwargs):
+        self.auth = kwargs["auth"]
+
+    def onInit(self):
+        status, resp = self.auth.get_device_code()
+        if status == self.auth.ERROR:
+            label = self.getControl(9111)
+            label.setLabel(
+                "Устройство не поддерживается.\n"
+                "Посетите http://kino.pub/device/support для уточнения деталей."
+            )
+            return
+        label = self.getControl(9112)
+        label.setLabel(resp["verification_uri"].encode("utf-8"))
+        label = self.getControl(9113)
+        label.setLabel(resp["user_code"].encode("utf-8"))
+        self.auth.verify_device_code(int(resp["interval"]))
+
+    def onAction(self, action):
+        actionID = action.getId()
+        if actionID in [9, 10, 13, 14, 15]:
+            self.closeWindow()
+
+    def closeWindow(self):
+        self.close()
 
 
 class Auth(object):
+    settings = __settings__
     terminated = False
     timer = 0
     CLIENT_ID = "xbmc"
@@ -17,10 +52,8 @@ class Auth(object):
     OAUTH_API_URL = "http://api.service-kp.com/oauth2/device"
     ERROR, PENDING_STATUS, SUCCESS, EXPIRED = range(4)
 
-    def __init__(self, settings, window=None, afterAuth=None):
-        self.window = window
-        self.settings = settings
-        self.afterAuth = afterAuth
+    def __init__(self):
+        self.window = AuthWindow("auth.xml", _ADDON_PATH, __skinsdir__, auth=self)
 
     def close(self):
         if self.window is not None:
@@ -63,6 +96,7 @@ class Auth(object):
     def reauth(self):
         self.access_token = ""
         self.device_token = ""
+        self.showActivationWindow()
 
     def request(self, url, data):
         xbmc.log("REQUEST URL={}".format(url))
@@ -143,48 +177,19 @@ class Auth(object):
             return self.SUCCESS, resp
         return self.ERROR, resp
 
-    def verify_device_code(self, interval, parent):
-        while not parent.stopped.wait(interval):
+    def verify_device_code(self, interval):
+        while True:
             success, resp = self.get_token()
             if success == self.SUCCESS:
-                self.afterAuth(force=True)
-                parent.closeWindow()
-                return True
+                update_device_info(force=True)
+                self.window.closeWindow()
+                break
+            time.sleep(interval)
+
+    def showActivationWindow(self):
+        xbmc.log("{}: No access_token. Show modal auth".format(__plugin__))
+        self.window.doModal()
+        xbmc.log("{}: Close modal auth".format(__plugin__))
 
 
-class AuthWindow(xbmcgui.WindowXMLDialog):
-    def __init__(self, *args, **kwargs):
-        self.stopped = threading.Event()
-        self.auth = Auth(kwargs["settings"], window=self, afterAuth=kwargs["afterAuth"])
-
-    def onInit(self):
-        status, resp = self.auth.get_device_code()
-        if status == self.auth.ERROR:
-            label = self.getControl(9111)
-            label.setLabel(
-                "Устройство не поддерживается.\n"
-                "Посетите http://kino.pub/device/support для уточнения деталей."
-            )
-            return
-        label = self.getControl(9112)
-        label.setLabel(resp["verification_uri"].encode("utf-8"))
-        label = self.getControl(9113)
-        label.setLabel(resp["user_code"].encode("utf-8"))
-        t = threading.Thread(
-            target=self.auth.verify_device_code,
-            args=[int(resp["interval"]), self]
-        )
-        t.daemon = True
-        t.start()
-
-    def onAction(self, action):
-        actionID = action.getId()
-        if actionID in [9, 10, 13, 14, 15]:
-            self.closeWindow()
-
-    def closeWindow(self):
-        self.stopped.set()
-        for thread in threading.enumerate():
-            if (thread.isAlive() and thread != threading.currentThread()):
-                thread.join(1)
-        self.close()
+auth = Auth()
