+++ video.kino.pub/resources/lib/data.py
@@ -1,0 +1,8 @@
+import xbmcaddon
+
+__id__ = "video.kino.pub"
+__addon__ = xbmcaddon.Addon(id=__id__)
+__settings__ = xbmcaddon.Addon(id=__id__)
+__skinsdir__ = "DefaultSkin"
+__language__ = __addon__.getLocalizedString
+__plugin__ = "plugin://{}".format(__id__)
